+++ src/Utility/System.cpp
@@ -115,7 +115,7 @@ void System::initSystemInfo()
 
 
 
-	Uint32 start, now, totalStart, totalNow;
+	Uint64 start, now, totalStart, totalNow;
 	start = SDL_GetPerformanceCounter();
 	totalStart = SDL_GetPerformanceCounter();
